Purge more files

+++ code/ui/generalhud/fullscreenhintmenu/FullScreenHintMenu.cs
@@ -1,55 +1,54 @@
 using Sandbox.UI;
 
-namespace TTT.UI
+namespace TTT.UI;
+
+public class FullScreenHintMenu : Panel
 {
-	public class FullScreenHintMenu : Panel
+	public static FullScreenHintMenu Instance;
+	public Panel ActivePanel { get; private set; }
+	public bool IsForcedOpen { get; set; } = false;
+
+	public FullScreenHintMenu()
 	{
-		public static FullScreenHintMenu Instance;
-		public Panel ActivePanel { get; private set; }
-		public bool IsForcedOpen { get; set; } = false;
+		Instance = this;
 
-		public FullScreenHintMenu()
+		StyleSheet.Load( "/ui/generalhud/fullscreenHintmenu/FullScreenHintMenu.scss" );
+
+		AddClass( "background-color-secondary" );
+		AddClass( "fullscreen" );
+
+		this.Style.ZIndex = 2;
+
+		this.EnableFade( false );
+	}
+
+	public void ForceOpen( Panel panel )
+	{
+		IsForcedOpen = true;
+		ActivePanel = null;
+
+		Open( panel );
+	}
+
+	public void Open( Panel panel )
+	{
+		if ( ActivePanel != null )
 		{
-			Instance = this;
-
-			StyleSheet.Load( "/ui/generalhud/fullscreenHintmenu/FullScreenHintMenu.scss" );
-
-			AddClass( "background-color-secondary" );
-			AddClass( "fullscreen" );
-
-			this.Style.ZIndex = 2;
-
-			this.EnableFade( false );
+			return;
 		}
 
-		public void ForceOpen( Panel panel )
-		{
-			IsForcedOpen = true;
-			ActivePanel = null;
+		DeleteChildren( true );
+		ActivePanel = panel;
+		AddChild( panel );
+		this.EnableFade( true );
+	}
 
-			Open( panel );
-		}
-
-		public void Open( Panel panel )
-		{
-			if ( ActivePanel != null )
-			{
-				return;
-			}
-
-			DeleteChildren( true );
-			ActivePanel = panel;
-			AddChild( panel );
-			this.EnableFade( true );
-		}
-
-		public void Close()
-		{
-			if ( ActivePanel == null || IsForcedOpen ) return;
-			this.EnableFade( false );
-			DeleteChildren( true );
-			ActivePanel = null;
-			IsForcedOpen = false;
-		}
+	public void Close()
+	{
+		if ( ActivePanel == null || IsForcedOpen ) return;
+		this.EnableFade( false );
+		DeleteChildren( true );
+		ActivePanel = null;
+		IsForcedOpen = false;
 	}
 }
